Unify remote_server crash handling
