+++ crates/crashes/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "crashes"
+version = "0.1.0"
+publish.workspace = true
+edition.workspace = true
+license = "GPL-3.0-or-later"
+
+[dependencies]
+crash-handler.workspace = true
+log.workspace = true
+minidumper.workspace = true
+paths.workspace = true
+smol.workspace = true
+uuid.workspace = true
+zlog.workspace = true
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/crashes.rs"
