+++ crates/remote_server/Cargo.toml
@@ -67,8 +67,11 @@ watch.workspace = true
 worktree.workspace = true
 
 [target.'cfg(not(windows))'.dependencies]
+crashes.workspace = true
+crash-handler.workspace = true
 fork.workspace = true
 libc.workspace = true
+minidumper.workspace = true
 
 [dev-dependencies]
 assistant_tool.workspace = true
